net: shell: add MQTT shell
This adds the MQTT shell with the commands listed below. Every command
call its underlying function from the MQTT API (i.e. the command connect
calls the function mqtt_connect()):
 - connect (call mqtt_connect())
 - ping (call mqtt_ping())
 - disconnect (call mqtt_connect())
 - publish (call mqtt_publish())
 - abort (call mqtt_abort())
 - live (call mqtt_live())
 - keepalive-time-left (call mqtt_keepalive_time_left())

The code is taken from mqtt_azure[1], thefore, the MQTT connection
address/port, client-id and credentials are via the following Kconfig
symbols:
 - MQTT_SHELL_CLIENT_ID (string)
 - MQTT_SHELL_ADDR (IPv4's addr string)
 - MQTT_SHELL_PORT (IPv4's port)
 - MQTT_SHELL_USER_NAME (string)
 - MQTT_SHELL_PASSWORD (string)

Note: The TCP/TLS connection and the subscribe/unsubscribe commands are
not implemented yet.

Important: The shell does not maintain the MQTT connection. As a
consequence, the command live has to be run whenever the counter
keep-alive reaches 0, or the command ping has to be run before the
counter reaches 0 to maintain the connection. Publishing a message or
pinging after the keepalive closes the connection and it has to be
reconnected. All the three commands publish, ping and live reset the
keep-alive counter.

[1]: https://github.com/zephyrproject-rtos/zephyr/tree/zephyr-v3.5.0/samples/net/cloud/mqtt_azure